requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.4
pyyaml==6.0.2
python-dotenv==1.0.1
pytest==8.3.4
google-cloud-aiplatform==1.132.0
google-cloud-bigquery==3.26.0
google-cloud-storage==3.0.0
# cloud-sql-python-connector==1.13.0
